+++ src/logic/product/Orderable.java
@@ -3,7 +3,7 @@ package logic.product;
 public interface Orderable {
     
     public abstract Orderable clone();
-    public abstract double getPrice();
+    public abstract int getPrice();
     public abstract String getDetail();
     public abstract void add(Orderable element);
     public abstract Orderable remove(int index);
